+++ clients/spot/tests/unit/websocket-api/TradeApiTest.test.ts
@@ -60,6 +60,7 @@ import {
 } from '../../../src/websocket-api';
 import {
     OpenOrdersCancelAllRequest,
+    OrderAmendKeepPriorityRequest,
     OrderCancelRequest,
     OrderCancelReplaceRequest,
     OrderListCancelRequest,
@@ -341,6 +342,234 @@ describe('TradeApi', () => {
                 try {
                     websocketAPIClient = new TradeApi(websocketBase);
                     const responsePromise = websocketAPIClient.openOrdersCancelAll(params);
+                    await expect(responsePromise).rejects.toThrow(/^Request timeout for id:/);
+                    resolveTest(true);
+                } catch (error) {
+                    resolveTest(error);
+                }
+            });
+            mockWs.emit('open');
+
+            const result = await testComplete;
+            if (result instanceof Error) {
+                throw result;
+            }
+        }, 10000);
+    });
+
+    describe('orderAmendKeepPriority()', () => {
+        beforeEach(async () => {
+            mockWs = Object.assign(new EventEmitter(), {
+                close: jest.fn(),
+                ping: jest.fn(),
+                pong: jest.fn(),
+                send: jest.fn(),
+                readyState: WebSocketClient.OPEN,
+                OPEN: WebSocket.OPEN,
+                CLOSED: WebSocket.CLOSED,
+            }) as unknown as jest.Mocked<WebSocketClient> & EventEmitter;
+
+            (WebSocketClient as jest.MockedClass<typeof WebSocketClient>).mockImplementation(
+                () => mockWs
+            );
+
+            const config = new ConfigurationWebsocketAPI({
+                apiKey: 'test-api-key',
+                apiSecret: 'test-api-secret',
+                wsURL: 'ws://localhost:3000',
+                timeout: 1000,
+            });
+
+            websocketBase = new WebsocketAPIBase(config);
+            websocketBase.connect();
+        });
+
+        afterEach(async () => {
+            if (websocketBase) {
+                await websocketBase.disconnect();
+            }
+            jest.clearAllMocks();
+            jest.clearAllTimers();
+        });
+
+        it('should execute orderAmendKeepPriority() successfully', async () => {
+            mockResponse = JSONParse(
+                JSONStringify({
+                    id: '56374b46-3061-486b-a311-89ee972eb648',
+                    status: 200,
+                    result: {
+                        transactTime: 1741924229819,
+                        executionId: 60,
+                        amendedOrder: {
+                            symbol: 'BTUCSDT',
+                            orderId: 23,
+                            orderListId: 4,
+                            origClientOrderId: 'my_pending_order',
+                            clientOrderId: 'xbxXh5SSwaHS7oUEOCI88B',
+                            price: '1.00000000',
+                            qty: '5.00000000',
+                            executedQty: '0.00000000',
+                            preventedQty: '0.00000000',
+                            quoteOrderQty: '0.00000000',
+                            cumulativeQuoteQty: '0.00000000',
+                            status: 'NEW',
+                            timeInForce: 'GTC',
+                            type: 'LIMIT',
+                            side: 'BUY',
+                            workingTime: 1741924204920,
+                            selfTradePreventionMode: 'NONE',
+                        },
+                        listStatus: {
+                            orderListId: 4,
+                            contingencyType: 'OTO',
+                            listOrderStatus: 'EXECUTING',
+                            listClientOrderId: '8nOGLLawudj1QoOiwbroRH',
+                            symbol: 'BTCUSDT',
+                            orders: [
+                                {
+                                    symbol: 'BTCUSDT',
+                                    orderId: 23,
+                                    clientOrderId: 'xbxXh5SSwaHS7oUEOCI88B',
+                                },
+                                {
+                                    symbol: 'BTCUSDT',
+                                    orderId: 22,
+                                    clientOrderId: 'g04EWsjaackzedjC9wRkWD',
+                                },
+                                {
+                                    symbol: 'BTCUSDT',
+                                    orderId: 23,
+                                    clientOrderId: 'xbxXh5SSwaHS7oUEOCI88B',
+                                },
+                                {
+                                    symbol: 'BTCUSDT',
+                                    orderId: 22,
+                                    clientOrderId: 'g04EWsjaackzedjC9wRkWD',
+                                },
+                            ],
+                        },
+                    },
+                    rateLimits: [
+                        {
+                            rateLimitType: 'REQUEST_WEIGHT',
+                            interval: 'MINUTE',
+                            intervalNum: 1,
+                            limit: 6000,
+                            count: 1,
+                        },
+                    ],
+                })
+            );
+            mockResponse.id = randomString();
+
+            const params: OrderAmendKeepPriorityRequest = {
+                symbol: 'BNBUSDT',
+                newQty: 1.0,
+            };
+
+            let resolveTest: (value: unknown) => void;
+            const testComplete = new Promise((resolve) => {
+                resolveTest = resolve;
+            });
+
+            websocketBase.on('open', async (conn: WebsocketAPIBase) => {
+                try {
+                    websocketAPIClient = new TradeApi(conn);
+                    const sendMsgSpy = jest.spyOn(conn, 'sendMessage');
+                    const responsePromise = websocketAPIClient.orderAmendKeepPriority({
+                        id: mockResponse?.id,
+                        ...params,
+                    });
+                    mockWs.emit('message', JSONStringify(mockResponse));
+                    const response = await responsePromise;
+                    expect(response.data).toEqual(mockResponse.result ?? mockResponse.response);
+                    expect(response.rateLimits).toEqual(mockResponse.rateLimits);
+                    expect(sendMsgSpy).toHaveBeenCalledWith(
+                        '/order.amend.keepPriority'.slice(1),
+                        params,
+                        { isSigned: true, withApiKey: false }
+                    );
+                    resolveTest(true);
+                } catch (error) {
+                    resolveTest(error);
+                }
+            });
+            mockWs.emit('open');
+
+            const result = await testComplete;
+            if (result instanceof Error) {
+                throw result;
+            }
+        });
+
+        it('should handle server error responses gracefully', async () => {
+            mockResponse = {
+                id: randomString(),
+                status: 400,
+                error: {
+                    code: -2010,
+                    msg: 'Account has insufficient balance for requested action.',
+                },
+                rateLimits: [
+                    {
+                        rateLimitType: 'ORDERS',
+                        interval: 'SECOND',
+                        intervalNum: 10,
+                        limit: 50,
+                        count: 13,
+                    },
+                ],
+            };
+
+            const params: OrderAmendKeepPriorityRequest = {
+                symbol: 'BNBUSDT',
+                newQty: 1.0,
+            };
+
+            let resolveTest: (value: unknown) => void;
+            const testComplete = new Promise((resolve) => {
+                resolveTest = resolve;
+            });
+
+            websocketBase.on('open', async (conn: WebsocketAPIBase) => {
+                try {
+                    websocketAPIClient = new TradeApi(conn);
+                    const responsePromise = websocketAPIClient.orderAmendKeepPriority({
+                        id: mockResponse?.id,
+                        ...params,
+                    });
+                    mockWs.emit('message', JSONStringify(mockResponse));
+                    await expect(responsePromise).rejects.toMatchObject(mockResponse.error!);
+                    resolveTest(true);
+                } catch (error) {
+                    resolveTest(error);
+                }
+            });
+            mockWs.emit('open');
+
+            const result = await testComplete;
+            if (result instanceof Error) {
+                throw result;
+            }
+        });
+
+        it('should handle request timeout gracefully', async () => {
+            jest.useRealTimers();
+
+            const params: OrderAmendKeepPriorityRequest = {
+                symbol: 'BNBUSDT',
+                newQty: 1.0,
+            };
+
+            let resolveTest: (value: unknown) => void;
+            const testComplete = new Promise((resolve) => {
+                resolveTest = resolve;
+            });
+
+            websocketBase.on('open', async (conn: WebsocketAPIBase) => {
+                try {
+                    websocketAPIClient = new TradeApi(websocketBase);
+                    const responsePromise = websocketAPIClient.orderAmendKeepPriority(params);
                     await expect(responsePromise).rejects.toThrow(/^Request timeout for id:/);
                     resolveTest(true);
                 } catch (error) {